Client carpool modal

+++ client/src/components/Carpool.vue
@@ -1,16 +1,16 @@
 <template>
   <div class="card-container">
-    <div class="w-100" @click="dropdown = !dropdown">
+    <div class="w-100 title-container" @click="dropdown = !dropdown">
       <br />
       <div class="d-flex justify-content-center align-items-center">
         <h4 class="mb-0">Carpool</h4>
         <span class="badge badge-primary badge-pill ml-1">{{carpools.length}}</span>
       </div>
       <br />
-      <div class="arrow" v-if="!dropdown">
+      <div class="arrow down-arrow" v-if="!dropdown">
         <i class="fas fa-angle-double-down"></i>
       </div>
-      <div v-else class="arrow">
+      <div v-else class="arrow up-arrow">
         <i class="fas fa-angle-double-up"></i>
       </div>
     </div>
@@ -18,7 +18,7 @@
     <div v-if="dropdown" class="dropdown pb-3">
       <div v-for="carpool in carpools" :key="carpool._id" class="d-inline-block">
         <div class="p-3">
-          <h6 class="car-name bg-lightest-grey mb-0 pt-2">
+          <h6 class="car-name bg-lightest-grey mb-0 pt-2 pl-1 pr-1">
             {{ carpool.name }}
             <i
               @click="removeCarpool(carpool._id)"
@@ -54,34 +54,21 @@
       <hr />
       <h5 class="pb-2">New carpool</h5>
       <div class="col-12">
-        <form @submit.prevent="createCarpool()" class="row carpool-form d-flex direction-column">
-          <label for="carpool-name" class="col-6 col-md-2 pr-1">Carpool name</label>
-          <input
-            type="text"
-            id="carpool-name"
-            class="col-6 col-md-3"
-            v-model="newCarpool.name"
-            placeholder="Enter name"
-            required
-          />
-
-          <label for="carpool-total-seats" class="col-6 col-md-2 pl-2 pr-1">Number of seats</label>
-          <input
-            type="number"
-            min="1"
-            class="col-6 col-md-3"
-            v-model="newCarpool.totalSeats"
-            placeholder="Total number of seats"
-            required
-          />
-          <button class="btn-dark mx-auto text-light-grey" type="submit">Add</button>
-        </form>
+        <button @click="createCarpool()" class="btn-dark mx-auto text-light-grey">Add</button>
       </div>
+    </div>
+    <div
+      v-if="dropdown"
+      @click="dropdown = !dropdown"
+      class="w-100 text-right arrow bottom-up-arrow"
+    >
+      <i class="fas fa-angle-double-up mr-3"></i>
     </div>
   </div>
 </template>
 
 <script>
+import Swal from "sweetalert2";
 import NotificationService from "@/NotificationService";
 export default {
   name: "Carpool",
@@ -92,27 +79,18 @@ export default {
   },
   data() {
     return {
-      newCarpool: {
-        name: "",
-        totalSeats: 0,
-        tripId: this.$route.params.tripId,
-        tripAuthorId: this.tripData.authorId
-      },
       dropdown: false
     };
   },
   methods: {
-    createCarpool() {
-      let carpool = { ...this.newCarpool };
-      console.log(carpool);
-      let tripId = this.$route.params.tripId;
-      this.$store.dispatch("addCarpool", { tripId, carpool });
-      this.newCarpool = {
-        name: "",
-        totalSeats: 0,
+    async createCarpool() {
+      let data = await NotificationService.inputCar("Car Details");
+      let carpool = {
+        ...data,
         tripId: this.$route.params.tripId,
         tripAuthorId: this.tripData.authorId
       };
+      this.$store.dispatch("addCarpool", { tripId: carpool.tripId, carpool });
     },
     addOccupant(tripId, carpoolId) {
       console.log("carpool id from add occupant", carpoolId);
@@ -121,8 +99,6 @@ export default {
     },
     removeOccupant(carpoolId, occupantId, occupantProfileId, carpoolAuthors) {
       let tripId = this.$route.params.tripId;
-
-      console.log(this.$store.state.profile._id, occupantProfileId);
       if (
         this.$store.state.profile._id == occupantProfileId ||
         carpoolAuthors.includes(this.$store.state.user._id)
@@ -137,10 +113,21 @@ export default {
           "You can't remove this person from a carpool"
         );
     },
-
     removeCarpool(carpoolId) {
-      let tripId = this.$route.params.tripId;
-      this.$store.dispatch("removeCarpool", { tripId, carpoolId });
+      Swal.fire({
+        title: "Are You Sure You Want To Delete This Carpool?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          let tripId = this.$route.params.tripId;
+          this.$store.dispatch("removeCarpool", { tripId, carpoolId });
+          Swal.fire("Deleted!", "Your carpool has been deleted.", "success");
+        }
+      });
     }
   },
   computed: {
@@ -166,29 +153,31 @@ export default {
   align-items: center;
   margin-top: 1em;
   min-height: 5em;
-  cursor: pointer;
   margin: 0px 0px;
+}
+.title-container {
+  cursor: pointer;
+  position: relative;
 }
 .top-card {
   background: rgba(255, 162, 75, 0.75) !important;
 }
 .arrow {
   font-size: 1.5em;
-  position: absolute;
-  right: 5%;
-  bottom: 1%;
 }
 .down-arrow {
   position: absolute;
-  right: 5%;
-  bottom: 1%;
+  right: 2%;
+  top: 1%;
 }
 .up-arrow {
   position: absolute;
   right: 2%;
-  top: 0;
+  top: 1%;
 }
-
+.bottom-up-arrow {
+  cursor: pointer;
+}
 .car-name {
   max-width: 10em;
 }
